+++ packages/config-utils/standalone/services/default/entitlements.js
