+++ packages/config-utils/standalone/services/index.js
@@ -1,0 +1,10 @@
+const defaultServices = require('./default');
+const rbac = require('./rbac');
+const backofficeProxy = require('./backofficeProxy');
+
+module.exports = {
+  defaultServices,
+  rbac,
+  backofficeProxy
+};
+
